升级到angular-cli@1.0.0-rc2

+++ tasks/build.js
@@ -1,29 +1,43 @@
 "use strict";
-var rimraf = require('rimraf');
-var path = require('path');
-var Task = require('../ember-cli/lib/models/task');
-var webpack = require('webpack');
-var webpack_config_1 = require('../models/webpack-config');
-var utils_1 = require('../models/webpack-configs/utils');
-var config_1 = require('../models/config');
 Object.defineProperty(exports, "__esModule", { value: true });
+const rimraf = require("rimraf");
+const path = require("path");
+const webpack = require("webpack");
+const app_utils_1 = require("../utilities/app-utils");
+const webpack_config_1 = require("../models/webpack-config");
+const utils_1 = require("../models/webpack-configs/utils");
+const config_1 = require("../models/config");
+const fs = require('fs');
+const Task = require('../ember-cli/lib/models/task');
+const SilentError = require('silent-error');
 exports.default = Task.extend({
     run: function (runTaskOptions) {
-        var _this = this;
-        var project = this.cliProject;
-        var outputPath = runTaskOptions.outputPath || config_1.CliConfig.fromProject().config.apps[0].outDir;
+        const project = this.cliProject;
+        const config = config_1.CliConfig.fromProject().config;
+        const app = app_utils_1.getAppFromConfig(runTaskOptions.app);
+        const outputPath = runTaskOptions.outputPath || app.outDir;
+        if (project.root === outputPath) {
+            throw new SilentError('Output path MUST not be project root directory!');
+        }
+        if (config.project && config.project.ejected) {
+            throw new SilentError('An ejected project cannot use the build command anymore.');
+        }
         rimraf.sync(path.resolve(project.root, outputPath));
-        var webpackConfig = new webpack_config_1.NgCliWebpackConfig(runTaskOptions).config;
-        var webpackCompiler = webpack(webpackConfig);
-        var statsConfig = utils_1.getWebpackStatsConfig(runTaskOptions.verbose);
-        return new Promise(function (resolve, reject) {
-            var callback = function (err, stats) {
+        const webpackConfig = new webpack_config_1.NgCliWebpackConfig(runTaskOptions, app).buildConfig();
+        const webpackCompiler = webpack(webpackConfig);
+        const statsConfig = utils_1.getWebpackStatsConfig(runTaskOptions.verbose);
+        return new Promise((resolve, reject) => {
+            const callback = (err, stats) => {
                 if (err) {
                     return reject(err);
                 }
-                _this.ui.writeLine(stats.toString(statsConfig));
+                this.ui.writeLine(stats.toString(statsConfig));
                 if (runTaskOptions.watch) {
                     return;
+                }
+                if (!runTaskOptions.watch && runTaskOptions.statsJson) {
+                    const jsonStats = stats.toJson('verbose');
+                    fs.writeFileSync(path.resolve(project.root, outputPath, 'stats.json'), JSON.stringify(jsonStats, null, 2));
                 }
                 if (stats.hasErrors()) {
                     reject();
@@ -33,18 +47,18 @@ exports.default = Task.extend({
                 }
             };
             if (runTaskOptions.watch) {
-                webpackCompiler.watch({}, callback);
+                webpackCompiler.watch({ poll: runTaskOptions.poll }, callback);
             }
             else {
                 webpackCompiler.run(callback);
             }
         })
-            .catch(function (err) {
+            .catch((err) => {
             if (err) {
-                _this.ui.writeError('\nAn error occured during the build:\n' + ((err && err.stack) || err));
+                this.ui.writeError('\nAn error occured during the build:\n' + ((err && err.stack) || err));
             }
             throw err;
         });
     }
 });
-//# sourceMappingURL=/Users/twer/dev/sdk/angular-cli/packages/@angular/cli/tasks/build.js.map
+//# sourceMappingURL=/users/twer/private/gde/angular-cli/tasks/build.js.map
